typescript/requirements.txt
pandas==2.3.3
pywebview==6.1
# pytest==9.0.2